+++ app/src/main/java/com/development/jaba/fragments/CarListFragment.java
@@ -3,6 +3,7 @@ package com.development.jaba.fragments;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.Fragment;
@@ -13,6 +14,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.afollestad.materialdialogs.DialogAction;
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.development.jaba.adapters.CarRowAdapter;
 import com.development.jaba.adapters.OnRecyclerItemClicked;
@@ -119,14 +121,15 @@ public class CarListFragment extends Fragment {
                     case 1:
                         DialogHelper.showYesNoDialog(String.format(getString(R.string.dialog_delete_car_title), mCarAdapter.getLastClickedItem().toString()),
                                 getText(R.string.dialog_delete_car_content),
-                                new MaterialDialog.ButtonCallback() {
+                                new MaterialDialog.SingleButtonCallback() {
                                     @Override
-                                    public void onPositive(MaterialDialog dialog) {
-                                        super.onPositive(dialog);
-                                        Car selectedCar = mCarAdapter.getItem(position);
-                                        mContext.deleteCar(selectedCar);
-                                        mCars.remove(selectedCar);
-                                        mCarAdapter.notifyItemRemoved(position);
+                                    public void onClick(@NonNull MaterialDialog materialDialog, @NonNull DialogAction dialogAction) {
+                                        if (dialogAction == DialogAction.POSITIVE) {
+                                            Car selectedCar = mCarAdapter.getItem(position);
+                                            mContext.deleteCar(selectedCar);
+                                            mCars.remove(selectedCar);
+                                            mCarAdapter.notifyItemRemoved(position);
+                                        }
                                     }
                                 },
                                 getActivity());
